Giorgi's Attempt
added two csv files that should have correspodning coordinates and the closest weather stations according to the data Marton provided for me as a txt file, as well as the python notebook, where I did the merging of the datasets. In the notebook I had to also import a json file that you can also see here because I kept getting NANs for coordinates and stations.